+++ CHANGELOG.md
@@ -18,3 +18,4 @@ used to track changes in this project.
 
 ### Changed
 - Migrate to UUIDs for primary keys
+- LoginWithTokenQuery to LoginWithTokenCommand